+++ SimpleGUICS2Pygame/simplegui_lib_draw.py
@@ -11,10 +11,8 @@ https://bitbucket.org/OPiMedia/simpleguics2pygame
 
 :license: GPLv3 --- Copyright (C) 2013, 2015, 2020 Olivier Pirson
 :author: Olivier Pirson --- http://www.opimedia.be/
-:version: March 14, 2020
+:version: March 25, 2020
 """
-
-from __future__ import print_function
 
 # print('IMPORT', __name__)