+++ SimpleGUICS2Pygame/test/test_arc.py
@@ -2,15 +2,17 @@
 # -*- coding: latin-1 -*-
 
 """
-Test draw circles.
+Test draw arcs.
 
 Piece of SimpleGUICS2Pygame.
 https://bitbucket.org/OPiMedia/simpleguics2pygame
 
-:license: GPLv3 --- Copyright (C) 2013, 2020 Olivier Pirson
+:license: GPLv3 --- Copyright (C) 2020 Olivier Pirson
 :author: Olivier Pirson --- http://www.opimedia.be/
 :version: March 26, 2020
 """
+
+import math
 
 try:
     import simplegui
@@ -37,8 +39,7 @@ else:
     PYGAME_VERSION = ''
     GUI_VERSION = 'simplegui'
 
-
-TEST = 'test circle'
+TEST = 'test arc'
 
 WIDTH = 400
 HEIGHT = 200
@@ -46,7 +47,7 @@ HEIGHT = 200
 
 def draw(canvas):
     """
-    Draw concentric circles.
+    Draw concentric arcs.
 
     :param canvas: simpleguics2pygame.Canvas or simplegui.Canvas
     """
@@ -55,20 +56,36 @@ def draw(canvas):
 
     line_width = 11
 
-    for radius in range(1, WIDTH, 30):
-        canvas.draw_circle((WIDTH - 1 - x, y), radius, line_width,
-                           'rgba(0,255,0,.5)')
-        canvas.draw_circle((WIDTH - 1 - x, y), radius, 1, 'Lime')
+    for radius in range(1, WIDTH, 20):
+        start_angle = radius // 20
+        end_angle = radius // 20 + math.pi
+        canvas.draw_arc((WIDTH - 1 - x, y), radius,
+                        start_angle, end_angle, line_width,
+                        'rgba(0,255,0,.5)')
+        canvas.draw_arc((WIDTH - 1 - x, y), radius,
+                        start_angle, end_angle, 1, 'Lime')
 
-    for radius in range(1, WIDTH, 30):
-        canvas.draw_circle((x, y), radius, line_width, 'rgba(255,0,0,.5)')
-        canvas.draw_circle((x, y), radius, 1, 'Red')
+    for radius in range(1, WIDTH, 20):
+        start_angle = radius // 20 + math.pi / 2
+        end_angle = radius // 20
+        canvas.draw_arc((x, y), radius,
+                        start_angle, end_angle, line_width, 'rgba(255,0,0,.5)')
+        canvas.draw_arc((x, y), radius,
+                        start_angle, end_angle, 1, 'Red')
 
     canvas.draw_circle((50, 50), 40, 10, 'Orange')
     canvas.draw_circle((110, 50), 40, 10, 'rgba(255,155,0,.5)')
 
     canvas.draw_circle((50, 110), 40, 5, 'rgba(0,155,255,.5)', 'Blue')
     canvas.draw_circle((110, 110), 40, 5, 'Aqua', 'rgba(0,255,255,.5)')
+
+    canvas.draw_arc((50, 50), 30, 0, math.pi, 10, 'Fuchsia')
+    canvas.draw_arc((50, 50), 30, math.pi * 5 / 4, math.pi * 7 / 4, 10, 'Red')
+    canvas.draw_arc((110, 50), 30, math.pi, math.pi * 5 / 2, 10,
+                    'rgba(255,0,255,.5)')
+
+    canvas.draw_arc((50, 110), 20, 0, 1, 10, 'rgba(255,0,0,.5)')
+    canvas.draw_arc((110, 110), 20, 1, 0, 10, 'rgba(255,0,0,.5)')
 
 
 #